// frontend/src/types/cv.ts
export interface ReferenceItem {
  name: string;
  title: string;
  company: string;
  contact: string;
}
export interface ReferencesContent {
  items: ReferenceItem[];
}